Sender match matches on Sender, from, reply_to.
Currently Sender only matches on 'from'. This resolves Issue #35
by making Sender match:
 - reply_to
 - from
 - sender
 - X-Original-Sender

This is similar to the functionality we have in `receiver`,
which matches 'to', 'cc', 'bcc', and 'list'.

We also speed up analysis by switching from getRawContent to
getHeader.

The MessageData.list now also includes the header 'List-ID' if it cannot
find a 'Mailing-list' header. This has been seen in a small sample of emails.