Allow rules with no actions.
Rows in the rules sheet can match a mail, but do not need to specify actions.
This is a different way of doing PR #29, instead of adding a new InboxActionType
we just decide to not worry about if a thread has any actions to do (as long as
it matched a rule).